Add service screen 2.0
The background needs to be changed and the email/password boxes need to
be better aligned in the center but there was an even number of items so
I wasn't sure how to do that. This should be at least fine for Wednesday
to show Icaro and I'll continue to update it to make it look better. If
you have suggestions I'd love to hear them.