Only 1 Cloud - Lights Are On
Dustin here. There is now only one cloud to go to the upload screen.
Also, the lights now go red and green when logged in/out. Dropbox works
correctly. Previously, only Dropbox had a method call to change the
log-in state. I changed that method but left the signature the same. I
just now added a method for Google and Skydrive so whoever has that
needs to call the setSkydriveLog(boolean mLoggedIn) method. Also, a
check login state needs to be made at the start of the app, not just
when someone hits log in or log out.