File Passing Addition
Ok, everything seems to work well for the demo. I added the file passing
code to DownloadFile.java so that it does not open the upload screen
unless it was able to succesfully download it. I also changed the move
files cloud button to open the filebrowser when clicked once and open
the upload screen if long clicked.

Let me know if there are any problems.