Fixed JAR Errors
Sean added a bunch of jars and they stilll had the paths to his desktop.
I copied the jars into the lib folder then used "Add Jars" to change the
path to the project\libs. Then I don't know if Sean got this error but
when I tried to run the project with the GSON 2.1 and Gson 1.7.1 and the
HTTPmime-4.0.3 and Httpmime-4.1.1 I got an error say they implemented
the same functions. So I deleted the ones with the older version number.
I just tested it and everything seemed to work