Alternative to uploadAll, error fixes, etc..
Fixed some file browser errors. I also set it up so that if you press
the back button your can go to the parent directory but I also made it
so that if you hold the back button you will go striaght back to the
home screen.

I don't believe I will be able to implement UploadAll by Friday.
However, I added the functinality of uploading a file multiple times. So
that you can upload to dropbox, box, and skydrive by dragging the icon
three times without having to go back and re-download it. I thought it
would be good for the demo. I also made it so that if you leave the
upload screen at all the file will be removed and set to null and it
will force you to go back.

I also added a remove funciton in uploadScreen so that unless you drag
the file to the sd card it will be removed when you exit the upload
screen. Not sure if it works but it doesn't matter as long as the upload
works for the demo.

If there are any errors or problems please let me know.