Error Fixing
I kept getting a funny error when I used the long back click to get back
to the main screen. I believe I fixed it. If anyone gets a "leaked
window" or "Illegal Argument exception" when they use the long back
click please let me know".

I also made it so that whenever the file broser is restarted either from
minimizing it or closing it out it will re-create. This will force the
user to re-populate each list making sure they are looking at the most
up-to-date version of the file browser.

Let me know if there are any problems!