Data Cap Warning file browsing, Fixed errors
Fixed errors when uploading to skydrive when not logged in. (Upload
Screen change)

Data Cap implement for file browsing. When you click on a file if your
WIFI is disconnected and the file is larger than 50 megabytes you will
get an alert dialog. I was not able to test this but I am pretty sure it
will work. Let me know if you have any problems.