Update: Fixed Errors and Minimum API 3.1
I am around until Saturday so I thought I would double check some stuff.

I found a few errors, fixed them.

I also did a little research and found that our actual minimum API
requirement is API(12) or  3.1 but in our Manifest file we said it was
API (15) or 4.0.3. So I changed the minimum SDK version to 12 and tested
it on an emulator running API 3.1 and everything worked. The GUI did not
look as good but that is too be expected with an older Android OS. I
tested it using the Nexus 7 and everything worked and looked pretty good
considering. This will allow us to reach more people.

If you want to change it back just change the line in Android Manifest
from 12 to 15.