Fixed an error when the move cloud button is long clicked
I never took out the code for when you long clicked the move cloud. So,
when you long click the move cloud it would bring you to the
uploadscreen and since sharefile was null you would get a
nullpointerexception.

I also put in an if statement when uploadscreen is created that if
sharefile is null that you will immedietely get the error "I don't know
what to upload" error and you get locked out of the screen. Just so that
if this every happens again for another reason you will get a toast
message and not a crash.

I tested it but let me know if there are any more issues. Thanks to
Brian for catching the error.