BIG FEATURE: Added a share menu option visible in top right corner of action bar
In UploadScreen.java is a some code regarding intents and what is
called a ShareActionProvider that shows a list of apps that can accept
the file type that is currently downloaded (email, sms, etc). It works
properly and shows the options, BUT I am unable to have the file
actually attach in the receiving application.

I believe the area to investigate is on line 250-255.