Skydrive Error Fix
I guess since Skydrive only has certain object types such as file,
folder, album etc. If a file is uploaded which somehow does not fit in
those speciifc types, the skydrive libraries will throw an Assertion
error if "Unknown Skydrive object type" when it attempts to create the
object. This fix skips the problematic file and displays all the other
files and folders.

Also, I do not know why but the .classpath changed as well. It looks
like just organization stuff that eclipse automatically did. I just left
it in.

I tested it and was able to successfully download and upload a picture.

If you find any problems let me know.