Working on making Beanstalk accept files from other apps.
Right now I'm working on adding the ability for any app to share TO
Beanstalk since we can now export to any other app. The way this works
is in the manifest we say that the app can accept files of particular
MIME types (audio, pictures, video, zip files, and MS docs) the fileUri
gets sent to SaveScreen activity which checks an SD card is mounted
then converts that fileUri to a new file. This new file is passed to
Upload screen as "savefile" and the upload screen activity starts.

The Uri is successfully sent right now but I'm having trouble making it
into a file that upload activity can recognize/use. For example, if you
go to the Gallery app and share an image, it does go to the upload
screen but when you drag the file is shows a generic file type (not a
picture) and all uploads fail. I'm sure this is related to how I'm
making a fileUri into a File.

In addition, this commit includes some beginning foundations for the
searching UI on save screen