Foundations for Search - not functional - remove if updating to 1.1.2 or until working
This commit includes the UI and files necessary for us to implement
search. The actual calls to do so just need to take place in
.SearchResultsActivity but I am not familiar with how to do this.
Should we decide to post another update before this works, be aware
that the UI for this should be commented out so users do not get
confused.

my references for these are available here:

http://developer.android.com/training/search/setup.html